requirements.txt
mdgen
Pillow
pycairo
seaborn
PyGObject
pandas
Jinja2
jsonschema
pyyaml
curlify
reportlab
matplotlib
setuptools
pytablewriter
axe_selenium_python
robotframework-faker
robotframework-browser
robotframework-requests
robotframework-jsonlibrary
robotframework-appiumlibrary
robotframework-seleniumlibrary
robotframework-screencaplibrary
robotframework-databaselibrary
robotframework-seleniumtestability